requirements.txt
Babel==2.9.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.4
click==7.1.2
dnspython==2.1.0
dominate==2.6.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-BabelEx==0.9.4
Flask-Bootstrap4==4.0.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.4
psycopg2-binary==2.8.6
pycparser==2.20
pytz==2020.5
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.22
visitor==0.1.3
Werkzeug==1.0.1
WTForm==1.0
WTForms==2.3.3
